@@ -13,7 +13,7 @@
     "source": "https://github.com/alcohol/iso3166"
   },
   "require": {
-    "php": ">=5.3"
+    "php": ">=5.4"
   },
   "autoload": {
     "psr-4": {
@@ -21,7 +21,7 @@
     }
   },
   "require-dev": {
-    "phpunit/phpunit": "~4.0",
+    "phpunit/phpunit": "~4.6",
     "phpunit/php-invoker": "@stable"
   },
   "autoload-dev": {
@@ -31,7 +31,7 @@
   },
   "extra": {
     "branch-alias": {
-      "dev-master": "2.0.x-dev"
+      "dev-master": "3.x-dev"
     }
   }
 }
